chatbot/requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.3
future==0.18.3
google-api-core==2.11.0
google-auth==2.16.1
google-cloud-core==2.3.2
google-cloud-translate==3.10.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
proto-plus==1.22.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
urllib3==1.26.14
webexteamssdk==1.6.1
Werkzeug==2.2.3
zipp==3.13.0
